improvements
- remove old json import files
- switch bundle themeing to JSON instead of inlined in JS
