+++ packages/main/src/json-imports/LocaleData.js
@@ -1,1 +1,2 @@
-import "@ui5/webcomponents-base/dist/cldr/CLDRAssets.js";
+import "@ui5/webcomponents-base/dist/json-imports/LocaleData.js";
+// the internal dependency to the actual LocaleData should be kept internal so it can be changed while apps still import this module
